+++ src/main/java/com/gildedgames/the_aether/client/renders/entity/DartBaseRenderer.java
@@ -60,19 +60,19 @@ public class DartBaseRenderer extends Render {
         GL11.glNormal3f(f10, 0.0F, 0.0F);
 
         tessellator.startDrawingQuads();
-        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, (double) f6, (double) f8);
-        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, (double) f7, (double) f8);
-        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, (double) f7, (double) f9);
-        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, (double) f6, (double) f9);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, f6, f8);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, f7, f8);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, f7, f9);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, f6, f9);
         tessellator.draw();
 
         GL11.glNormal3f(-f10, 0.0F, 0.0F);
 
         tessellator.startDrawingQuads();
-        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, (double) f6, (double) f8);
-        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, (double) f7, (double) f8);
-        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, (double) f7, (double) f9);
-        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, (double) f6, (double) f9);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, -2.0D, f6, f8);
+        tessellator.addVertexWithUV(-7.0D, 2.0D, 2.0D, f7, f8);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, 2.0D, f7, f9);
+        tessellator.addVertexWithUV(-7.0D, -2.0D, -2.0D, f6, f9);
         tessellator.draw();
 
         for (int var23 = 0; var23 < 5; ++var23) {
@@ -80,10 +80,10 @@ public class DartBaseRenderer extends Render {
             GL11.glNormal3f(0.0F, 0.0F, f10);
 
             tessellator.startDrawingQuads();
-            tessellator.addVertexWithUV(-8.0D, -2.0D, 0.0D, (double) f2, (double) f4);
-            tessellator.addVertexWithUV(8.0D, -2.0D, 0.0D, (double) f3, (double) f4);
-            tessellator.addVertexWithUV(8.0D, 2.0D, 0.0D, (double) f3, (double) f5);
-            tessellator.addVertexWithUV(-8.0D, 2.0D, 0.0D, (double) f2, (double) f5);
+            tessellator.addVertexWithUV(-8.0D, -2.0D, 0.0D, f2, f4);
+            tessellator.addVertexWithUV(8.0D, -2.0D, 0.0D, f3, f4);
+            tessellator.addVertexWithUV(8.0D, 2.0D, 0.0D, f3, f5);
+            tessellator.addVertexWithUV(-8.0D, 2.0D, 0.0D, f2, f5);
             tessellator.draw();
         }
